Create single page html containing all docs
Allow users to navigate to root/single-page-html/
to quickly ctrl-f through the documentation. There
is an issue with the includes command which causes
the first paragraph to be rendered as a code block.
As a result, an empty liquid block is placed right
after the front matter in every markdown file which
blocks this from happening